Better

+++ src/Processors/QueryPlan/DistributedCreateLocalPlan.h
@@ -3,6 +3,7 @@
 #include <Core/QueryProcessingStage.h>
 #include <Parsers/IAST_fwd.h>
 #include <Processors/QueryPlan/QueryPlan.h>
+#include <Processors/ResizeProcessor.h>
 #include <Storages/MergeTree/ParallelReplicasReadingCoordinator.h>
 
 namespace DB
@@ -13,8 +14,8 @@ std::unique_ptr<QueryPlan> createLocalPlan(
     const Block & header,
     ContextPtr context,
     QueryProcessingStage::Enum processed_stage,
-    UInt32 shard_num,
-    UInt32 shard_count,
+    size_t shard_num,
+    size_t shard_count,
     size_t replica_num,
     size_t replica_count,
     std::shared_ptr<ParallelReplicasReadingCoordinator> coordinator);
